move version command into cmd package
